Move interfaces to seperate file
now we have just over 340 lines in teh DraftPage.tsx file
