+++ client/src/components/DraftTool/DraftPage.tsx
@@ -7,36 +7,8 @@ import { handleBanPhase, handlePickPhase } from "./clientDraftHandler";
 
 import championsData from "./championRoles.json";
 import { DisplayBanImage, DisplayPickImage } from "./LoadChampLargeImages";
+import { Champion, DraftStateProps } from "./draftInterfaces";
 
-export interface Champion {
-  name: string;
-  roles: string[];
-}
-
-export interface DraftStateProps {
-  draftStarted: boolean;
-  activePhase:
-    | "banPhase1"
-    | "pickPhase1"
-    | "banPhase2"
-    | "pickPhase2"
-    | "finished"
-    | null;
-  phaseType: "pick" | "ban" | null;
-  blueUser: string;
-  redUser: string;
-  blueReady: boolean;
-  redReady: boolean;
-  timer: number;
-  bansArray: string[];
-  picksArray: string[];
-  banIndex: number;
-  pickIndex: number;
-  currentTurn: string;
-  displayTurn: "red" | "blue" | null;
-  bluePick: string | null;
-  redPick: string | null;
-}
 
 function DraftPage() {
   const [socket, setSocket] = useState<Socket | null>(null);
